more fb changes

+++ mooch-client/src/components/helpers/emailAuth.js
@@ -7,18 +7,13 @@ import {
 } from "firebase/auth";
 
 
-// userObject expected ---->
-// {
-//   email: "",
-//   password: "",
-//   fullName: "",
-// }
-
 const _apiUrl = "https://localhost:7082/api"
 
+//check our API to ensure that the firebase user that was just logged exists in our local SQL database
 const doesUserExist = (firebaseUserId) => {
-
-  return getToken().then((token) => fetch(`${_apiUrl}/UserExists/${firebaseUserId}`, {
+  
+  return getToken()
+    .then((token) => fetch(`${_apiUrl}/UserExists/${firebaseUserId}`, {
     method: "GET",
     headers: {
       Authorization: `Bearer ${token}`
@@ -27,6 +22,7 @@ const doesUserExist = (firebaseUserId) => {
   
 }
 
+//extract token from firebase response and return it here
 export const getToken = () => {
   const auth = getAuth();
   const currentUser = auth.currentUser;
@@ -42,31 +38,33 @@ export const emailAuth = {
   // Register New User
   register: function(userObj, navigate) {
     const auth = getAuth();
+    const userAuth = {};
     createUserWithEmailAndPassword(auth, userObj.email, userObj.password)
       .then((userCredential) => {
-        const auth = getAuth();
-        updateProfile(auth.currentUser, {
-          displayName: userObj.fullName,
-        }).then(
-          function() {
-            const userAuth = {
-              email: userCredential.user.email,
-              displayName: userObj.fullName,
-              uid: userCredential.user.uid,
-              type: "email",
-            };
-            // Saves the user to localstorage
-            localStorage.setItem("capstone_user", JSON.stringify(userAuth));
-            // Navigate us back to home
-  
+            
+              userAuth.email = userCredential.user.email;
+              userAuth.uid = userCredential.user.uid;
+              userAuth.type= "email";
+            doesUserExist(userCredential.user.uid)
+            .then((userExists) => {
+              if (!userExists)  {
+                  //navigate to new user page.
+                  navigate("/createuser")
+              } else {
+
+                // Saves the user to localstorage
+                localStorage.setItem("capstone_user", JSON.stringify(userAuth));
+                // Navigate us back to home
+                navigate("/")
+              }
+            })
           },
           function(error) {
             console.log("Email Register Name Error");
             console.log("error code", error.code);
             console.log("error message", error.message);
           }
-        );
-      })
+        )
       .catch((error) => {
         console.log("Email Register Error");
         console.log("error code", error.code);
@@ -77,7 +75,7 @@ export const emailAuth = {
   signIn: function(userObj, navigate) {
     return new Promise((res) => {
       const auth = getAuth();
-      let existingUser = {};
+      const existingUser = {};
       signInWithEmailAndPassword(auth, userObj.email, userObj.password)
         .then((SignInResponse) => {
           existingUser.email = SignInResponse.user.email;
@@ -89,8 +87,6 @@ export const emailAuth = {
             if (!userExists) {
               this.signOut();
             } else {
-  
-              
               // Saves the user to localstorage
               localStorage.setItem("capstone_user", JSON.stringify(existingUser));
               // Navigate us back to home
